rename de fattify et ajout de TODOs
LSLarose

+++ server/app/img-diff-generator/img-diff-generator/Image.cpp
@@ -27,20 +27,11 @@ void Image::setPixel(const unsigned & x, const unsigned & y, const Pixel & pixel
 ostream& operator<<(ostream & os, Image & image)
 {
 	vector<vector<Pixel>> pixels(image.getPixels());
-	for (int y = pixels[0].size() - 1; y >= 0; y--)
+	for (size_t y = pixels[0].size() - 1; y >= 0; y--)
 	{
 		for (unsigned x = 0; x < pixels.size(); x++)
 		{
 			Pixel pixel(pixels[x][y]);
-			//os << setw(3) << setfill('0')
-			//	<< (unsigned)pixel.r
-			//	<< ", ";
-			//os << setw(3) << setfill('0')
-			//	<< (unsigned)pixel.g
-			//	<< ", ";
-			//os << setw(3) << setfill('0')
-			//	<< (unsigned)pixel.b;
-			//os << "     ";
 			os << setw(2) << setfill('0') << hex << (unsigned)pixel.r
 				<< setw(2) << setfill('0') << hex << (unsigned)pixel.g
 				<< setw(2) << setfill('0') << hex << (unsigned)pixel.b
